Solve circular dependency
