+++ front/lib/api/files/mount_path.ts
@@ -1,9 +1,4 @@
-import type { Authenticator } from "@app/lib/auth";
-import type { FileResource } from "@app/lib/resources/file_resource";
-import { FileModel } from "@app/lib/resources/storage/models/files";
-import { isConversationFileUseCase } from "@app/types/files";
-
-// Path helpers for conversation file mount paths (gcsfuse sandbox mounting).
+// Pure path helpers for conversation file mount paths (gcsfuse sandbox mounting).
 
 export function getBaseMountPathForWorkspace({
   workspaceId,
@@ -23,7 +18,7 @@ export function getConversationFilesBasePath({
   return `${getBaseMountPathForWorkspace({ workspaceId })}conversations/${conversationId}/files/`;
 }
 
-function getConversationFilePath({
+export function getConversationFilePath({
   workspaceId,
   conversationId,
   fileName,
@@ -60,7 +55,7 @@ export function makeProcessedMountFileName(mountFilePath: string): string {
  * "report.pdf" + "fil_abc" → "report_fil_abc.pdf"
  * "Makefile" + "fil_abc" → "Makefile_fil_abc"
  */
-function disambiguateFileName(fileName: string, sId: string): string {
+export function disambiguateFileName(fileName: string, sId: string): string {
   const lastDot = fileName.lastIndexOf(".");
   if (lastDot <= 0) {
     return `${fileName}_${sId}`;
@@ -69,73 +64,4 @@ function disambiguateFileName(fileName: string, sId: string): string {
   const basename = fileName.substring(0, lastDot);
   const ext = fileName.substring(lastDot);
   return `${basename}_${sId}${ext}`;
-}
-
-/**
- * Single entry point for mount path resolution.
- *
- * Examines the file's use case and metadata to determine whether a mount path should be created.
- * Branches internally by use case:
- * - conversation / tool_output: mounts under w/{wId}/conversations/{cId}/files/
- * - (future use cases can be added here)
- *
- * No-ops if the file already has a mountFilePath or conditions aren't met.
- */
-export async function maybeResolveMountPath(
-  auth: Authenticator,
-  file: FileResource
-): Promise<void> {
-  // Skip if already resolved.
-  if (file.mountFilePath) {
-    return;
-  }
-
-  const { useCase, useCaseMetadata } = file;
-
-  if (isConversationFileUseCase(useCase) && useCaseMetadata?.conversationId) {
-    await resolveConversationMountPath(auth, file, {
-      conversationId: useCaseMetadata.conversationId,
-    });
-    return;
-  }
-
-  // TODO(2026-03-09 SANDBOX): Add support for project context.
-}
-
-/**
- * Resolve the mount path for a conversation file.
- *
- * Checks if the desired path is already taken via the unique index on mountFilePath. If taken,
- * disambiguates with the file's sId.
- *
- * Then calls file.setMountFilePath to persist and copy to GCS.
- */
-async function resolveConversationMountPath(
-  auth: Authenticator,
-  file: FileResource,
-  { conversationId }: { conversationId: string }
-): Promise<void> {
-  const owner = auth.getNonNullableWorkspace();
-
-  const desiredPath = getConversationFilePath({
-    workspaceId: owner.sId,
-    conversationId,
-    fileName: file.fileName,
-  });
-
-  // Check if the desired path is already taken (unique index lookup).
-  const existing = await FileModel.findOne({
-    attributes: ["id"],
-    where: { workspaceId: owner.id, mountFilePath: desiredPath },
-  });
-
-  const mountFilePath = existing
-    ? getConversationFilePath({
-        workspaceId: owner.sId,
-        conversationId,
-        fileName: disambiguateFileName(file.fileName, file.sId),
-      })
-    : desiredPath;
-
-  await file.setMountFilePath(auth, mountFilePath);
 }
